Eject floppy and installation ISO after the VM has been provisioned.
To work around an issue in the PV drivers for Windows, the floppy and
installation ISO were removed before restarting the guest. This was due
to the drivers bluescreening on a failed assert of the presence of a
floppy disk.

Now that behaviour has been fixed, the builder should behave in the
expected way, and not introduce an extra shutdown/restart.

Signed-off-by: Rob Dobson <rob.dobson@citrix.com>