Fix SSH handling in pool environment.
This adds a new build step executed after StepStartVmPaused. This sets
the "ssh_address" state variable to the IP of the host the VM was started on.
This enables SSH commands to work correctly in a pool environment.

This also modifies SSH calls to use this address rather than config.HostIp

Fixes #47